requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pillow==11.2.0
PyMuPDF==1.25.5
pytesseract==0.3.13
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.5.0
tqdm==4.67.1
Werkzeug==3.1.3
gunicorn
